Time: 0 ms (100%), Space: 8.9 MB (7.24%) - LeetHub

+++ 0050-powx-n/0050-powx-n.cpp
@@ -3,7 +3,7 @@ public:
     double myPow(double x, int n) {
         long double y=1;
         
-        y=pow(x,n);
+        y=(long double)pow(x,n);
         
         
     return y;
